all_versions.md
These download links are not working including e.g. if these could be updated and pointed to the correct resource page, thanks.
https://clis.ng.bluemix.net/download/bluemix-cli/latest/linux64

Customer is getting when clicking on linux64 e.g. - 
The page isn’t redirecting properly

Firefox has detected that the server is redirecting the request for this address in a way that will never complete.

    This problem can sometimes be caused by disabling or refusing to accept cookies.